HertaScannerBot/pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "discord-py>=2.4.0",
    "python-dotenv>=1.0.1",
    "selenium>=4.27.1",
    "tweepy>=4.15.0",
]
